chore: Move GIS to module, setup routes
