+++ api.planx.uk/modules/gis/service/local_authorities/metadata/camden.ts
